requirements.txt
pyrofork==2.3.24
motor==3.6.0
aiohttp==3.10.11
feedparser==6.0.11
TgCrypto
python-dotenv==1.0.1
